Split json and ndjson into separate reporters over a shared engine
The single jsonReporter branched on `format`. Split it into two
conditional-free reporters that each just say how a row is written:

- reporters/json.ts   — buffers rows into one array
- reporters/ndjson.ts — streams one row per line

The shared machinery (file discovery, row mapping, CI-gate tally, stdout
backpressure, flush-and-exit) moves into reporters/rowReporter.ts as
`runRowReporter(options, paths, { onRow, onDone })`, so neither reporter
duplicates the subtle gate/exit/backpressure logic. ReportRow + toReportRow (and
its tests) live there too. The mod.ts map now points straight at each reporter,
dropping the per-format closure wrappers.
